feat(plugins): installed telescope, changed colorscheme
This was a long time coming. I thought it would be a big plugin and this
increase "bloat", but working on real projects at my internship simply
using `:e **somefile` was taking ages to load.
Looking around a bit, it actually is really cool. More modules and
integrations to come.

Had to change colour scheme, as base-16 couldn't cope with telescope.
Configs now use the lua api for setting the colour scheme.
Also changed some strings to single quotes.