fix(leap): enable more elaborate colouring when jumping
This makes it a lot more clear when leap is triggered, and in what
direction it's going. Important to note, changing highlight groups
through an autocommand has to be done _before_ setting the colourscheme
for some reason. That's why I changed to order in which files get sources
in the root init.lua file.