feat(lsp): rework server configs
This is a kind of nasty stop gap solution until I better integrate my
Neovim config with my Nix config.

The benefit is I can split my server configs into separate configs. The
bad news is that it kind of breaks the "just works" part of
mason_lspconfig. I now have a mix of the following situations to deal
with:

- Server was installed with mason and uses just the defaults.
- Server was installed with mason and uses custom settings.
- Server was manually installed and uses just the defaults.
- Server was manually installed and uses custom settings.