feat(plugins): install nvim-orgmode
I've been experimenting with this, in particular to structure the peer
feedback and self evaluation grading system used at my study.

Many people swear by Org mode, and this is exactly the kind of nerdy
thing I would be into (you should see how I manage my finances :D), yet
I haven't fallen in love with it yet. People comment on the ease of
having Emacs style super-hyper-alt-meta-control-something-or-other to
toggle the state of a todo list item, but I maintain that the solid text
editing primitives of vim beat trying to turn your text files into a
typical hotkey powered GUI application. Org advertises itself as a plain
text file type, which is true enough, but using it outside of Emacs is
tough sell.

Folding and concealing is also not that big a deal to me. With
Telescope.nvim and oil.nvim I can find my way around files super
quickly, so I don't see a benefit of putting all info in a single file
and folding the parts you're not interested in.

I'm also curious about Neorg, as it's a native Neovim Org alternative
(though it also claims to be portable).

I really need an agenda, and a way to sync across devices. Remind is
awe-inspiring, but convoluted recurring events aren't really my use
case, more like one-off meetings and appointments and tasks.

Since Org is so highly regarded by its users, there has to be something
in it for me.