feat(lsp): customize `nil_ls` settings
In particular, disable autoArchive flake option. From my understanding,
this downloads flake inputs that aren't yet present in the store, for
instance when you've added a flake input but not yet rebuild the config.
I'd rather not that my LSP downloads stuff in the background, and I
don't want it asking every time.