feat: workaround to get rounded borders back around hover windows
nvim 0.11 introduced a global `border` option, and changed the way
borders are set for certain floating windows, like hover documentation.
As a result, hovers have no borders by default which looks pretty
terrible. But plugins like telescope are designed without a global
`border` option in mind, and look broken with it enabled. This is a
minor hack I spotted in a related issue on the telescope github repo
that wraps the hover function to temporarily enable rounded borders.