feat(plugins): install fancier virtual text diagnostics plugin
tiny-inline-diagnostics is a plugin that changes the appearance of
diagnostic virtual text. In particular, it makes it possible to break or
wrap long lines and showing multiple messages, while keeping it aligned
to the right outside of the main text. This makes the virtual text
actually readable, even in smaller windows, without displacing the main
text like the new nvim 0.11 `virtual_lines` option does.